archive/july-2025/requirements.txt
pipecatcloud
pipecat-ai[aws,daily,deepgram,openai,silero,webrtc,websocket]
pipecat-ai-small-webrtc-prebuilt
python-dotenv~=1.0.1
strands-agents